combining all harmonic osc into /examples
