+++ vmcnet/examples/__init__.py
@@ -1,0 +1,2 @@
+"""Examples for integration tests and pedagogy."""
+from . import harmonic_oscillator
