+++ vmcnet/models/__init__.py
@@ -1,4 +1,3 @@
 """Model building routines."""
 from . import antisymmetry
 from . import construct
-from . import harmonic_osc
